I've got the javascript running for the tablet-mode sidebar.
I've noticed some issues when resizing the window between the different widths I'm trying to account for. My idea so far is that I can fix these issues by restructuring the CSS so that the website is mobile-first.

To be honest, that doesn't seem like it should take too long, but revisiting code I thought I was done with is a pain.

Overall:

- Implement mobile screen sizes.

- Restructure to make mobile-first

- The javascript issue of toggling the sidebar and then changing the screen width could be counteracted by using vw as the units. Try it out and see.